enhance request dispatching by Content-Type
Add methods {Get,Post}As(...) which take an additional contentType
parameter, which is later on used when dispatching requests. If a
request has its Content-Type-Header explicitly set (which usually is
not the case) _only_ matching handlers are executed. If it isn't
set, any handler will do. This way, the current behaviour should be
backwards compatible with earlier versions.